project/requirements.txt
# OpenAI + Docling RAG System Dependencies

# --- OpenAI ---
openai>=1.0.0
langchain-openai>=0.1.0

# --- LangChain Core ---
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-qdrant>=0.1.0
langchain-text-splitters>=0.2.0
langgraph>=0.1.0
langgraph-checkpoint>=1.0.0
langsmith>=0.1.0

# --- Vector DB ---
qdrant-client>=1.9.0

# --- PDF Processing ---
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.5

# --- Docling OCR (for advanced PDF parsing) ---
docling>=2.0.0
docling-core>=2.0.0

# --- PyTorch (required by Docling) ---
# Note: On Linux, you may need: apt-get install libgl1-mesa-glx
torch>=2.2.0,<2.6.0

# --- OpenCLIP (for semantic image-query matching) ---
open-clip-torch>=2.24.0

# --- Web Framework ---
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# --- Utilities ---
pydantic>=2.0.0
pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.0.0
tqdm>=4.65.0
pyyaml>=6.0
rich>=13.0.0
tenacity>=8.2.0

# --- Data Processing ---
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
